Enable targeting of different devices for inferencing
Create an abstraction to handle all of the details of targeting different
devices.  Let users pick the device they want to use by exposing a new sub
command that tells the user what devices they are able to use based on their
setup.

When doing this detection work, be mindful of varying degrees of support from
different PyTorch versions and different PyTorch installations and different
hardware.  As a result, assume nothing is available and only use it if it is
confirmed to be available.

According to the following documentation:

https://github.com/pytorch/pytorch/blob/main/docs/source/notes/hip.rst#hip-interfaces-reuse-the-cuda-interfaces

AMD GPUs use `cuda` as the device name and use all of the same `torch.cuda`
interfaces.  It's a shame to see this as it further cements CUDA's staying
power, but that's how it is.  A special PyTorch installation is required to
enable this.

Intel GPUs require a special PyTorch installation as well where `torch.xpu`
interfaces are available.  Again, this is a shame to see as it further
fragments the PyTorch interface and it makes programming against a single
uniform interface hard (hence the need for this kind of an abstraction).  The
IPEX module should also be available and used on models after they are
transferred to the Intel GPU or for Intel CPUs to get the best performance.
Intel uses the `xpu` device name.

Apple Silicon GPU support was added in recent PyTorch versions, and is
indicated with the `mps` device name.

For the `cpu` device name, this is always available and is the default place
where models exist.  Use helpful packages to get detailed information about the
CPU when enumerating available devices.

And of course, the `cuda` device name is exactly what it is.  *sneers and
scorns*

The intended use is for model classes to take a DeviceManager rather than just
a device identifier and then use the corresponding stage_model and stage_cpu
functions to properly set everything up in whatever environment is necessary to
run inferencing on models.