Use current Intel API from IPEX
The package torch.xpu doesn't exist, as the sanity check documentation would
suggest.  Used the intel_extension_for_pytorch.xpu package instead.

The intel_extension_for_pytorch.xpu package is always installed, so judge
whether the xpu device is usable based on whether the
intel_extension_for_pytorch package is installed and if the xpu device is
valid.

If the module is available, always import it into the global namespace so that
all commands using that module in the device manager will work.